Add files via upload
A lot of changes on this commit. I got all of the text box styles complete and working. The really tough ones were the styles with glyphicons and the Dolla Billz style (which has fancy things like a tool tip and only allows numbers to be inserted into the text box).